Dates and isotope spreadsheet updates
I went in and added the last two batches of plant dates to the master spreadsheet. then, I re-exported the tabs and copied over to Git data folder for input.  I also changed the code so that I pull out both mammal and plant dates, but haven't finished cascading those changes through the rest of the code.