Update the plotting code
OK, Nate.  I think the plots are looking better.  I ended up scripting it so that all of the plots are done for you, cleaning up the labels and the axes.  You will see that there are more plots than are in your figure tables, so make sure you check the x and y labels to only pull the figures you need.  All of these plots are in "output/new sensitivity/..."